requirements.txt
# basic django and backend
django>=3.0.3,<4.0.0
python-memcached>=1.59,<2.0.0
psycopg2-binary>=2.8.4,<3.0.0

# excel export
openpyxl>=3.0.3,<4.0.0

# server
uwsgi>=2.0.18,<3.0.0

#django reversions
django-reversion>=3.0.5,<4.0.0
django-reversion-compare>=0.9.0,<2.0.0

# API access
djangorestframework>=3.11.0,<4.0.0